+++ frontend/src/app/pages/build-template-introduction/build-template-introduction.component.ts
@@ -5,7 +5,6 @@ import { Component, ViewEncapsulation } from '@angular/core';
   standalone: true,
   imports: [],
   templateUrl: './build-template-introduction.component.html',
-  styleUrl: './build-template-introduction.component.scss',
   encapsulation: ViewEncapsulation.None
 })
 export class BuildTemplateIntroductionComponent {}